requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
Django==4.1.5
django-cors-headers==4.3.1
django-countries==7.5.1
djangorestframework==3.15.1
gunicorn==21.2.0
mypy-extensions==0.4.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
psycopg2-binary==2.9.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
